Proyecto/scraping/scraping_becas/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Data Processing
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Database
sqlalchemy==2.0.23
mysqlclient==2.2.0
PyMySQL==1.1.0

# Web Interface
jinja2==3.1.2
aiofiles==23.2.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.0

# HTTP Client
httpx==0.25.2

# Data Validation
pydantic==2.5.0

# Browser Automation
webdriver-manager==4.0.1